+++ main/parsers.h
@@ -12,6 +12,14 @@
 #ifndef CTAGS_MAIN_PARSERS_H
 #define CTAGS_MAIN_PARSERS_H
 
+#ifdef HAVE_LIBXML
+#define XML_PARSER_LIST \
+	DbusIntrospectParser, \
+	GladeParser
+#else
+#define XML_PARSER_LIST
+#endif
+
 /* Add the name of any new parser definition function here */
 #define PARSER_LIST \
 	AdaParser, \